+++ integration/mocha-tests/tests/AudioTest.js
@@ -1,10 +1,9 @@
 const { describe, before, after, it } = require('mocha');
 const { v4: uuidv4 } = require('uuid');
 const { Window } = require('../utils/Window');
-const WebDriverFactory = require('../utils/WebDriverFactory');
-const SdkBaseTest = require('../utils/SdkBaseTest');
-const AudioTestPage = require('../pages/AudioTestPage');
-const { Logger, LogLevel } = require('../utils/Logger');
+
+const { LogLevel } = require('../utils/Logger');
+const setupTestEnvironment = require('../steps/TestSetup');
 
 /*
  * 1. Starts a meeting
@@ -16,66 +15,25 @@ const { Logger, LogLevel } = require('../utils/Logger');
  * 7. Checks if the other participant is not able to hear the audio
  * */
 describe('AudioTest', async function () {
-  let driverFactoryOne;
-  let baseTestConfigOne;
-  let pageOne;
   let test_window;
   let monitor_window;
   let test_attendee_id;
   let monitor_attendee_id;
   let meetingId;
   let numberOfSessions;
-  let failureCount;
-
-  before(async function () {
-    this.logger = new Logger('AudioTest');
-    this.logger.log('Retrieving the base test config');
-    baseTestConfigOne = new SdkBaseTest();
-    this.logger.log('Configuring the webdriver');
-    driverFactoryOne = new WebDriverFactory(
-      baseTestConfigOne.testName,
-      baseTestConfigOne.host,
-      baseTestConfigOne.testType,
-      baseTestConfigOne.url
-    );
-    await driverFactoryOne.build();
-    this.logger.log('Using the webdriver, opening the browser window');
-    numberOfSessions = driverFactoryOne.numberOfSessions;
-    driverFactoryOne.driver.manage().window().maximize();
-    this.logger.log('Instantiating selenium helper class');
-    pageOne = new AudioTestPage(driverFactoryOne.driver, this.logger);
-    failureCount = 0;
-  });
-
-  afterEach(async function () {
-    if (this.currentTest.state === 'failed') failureCount += 1;
-  });
-
-  after(async function () {
-    this.logger.log('Closing the webdriver');
-    const passed = failureCount === 0;
-
-    await driverFactoryOne.quit(passed);
-
-    if (passed === true) {
-      this.logger.log('AudioTest passed!!!', LogLevel.SUCCESS);
-      process.exit(0);
-    } else {
-      this.logger.log('AudioTest failed!!!', LogLevel.ERROR);
-      process.exit(1);
-    }
-  });
+  const testSetup = setupTestEnvironment('AudioTest');
+  testSetup.setupBaseTest();
 
   describe('on single session', async function () {
-    before(async function () {
-      if (numberOfSessions !== 1) {
-        this.logger.log(
-          `Skipping single session test because number of sessions required is ${numberOfSessions}`,
-          LogLevel.WARN
-        );
-        this.skip();
-      }
-    });
+    // before(async function () {
+    //   if (numberOfSessions !== 1) {
+    //     this.logger.log(
+    //       `Skipping single session test because number of sessions required is ${numberOfSessions}`,
+    //       LogLevel.WARN
+    //     );
+    //     this.skip();
+    //   }
+    // });
 
     afterEach(async function () {
       this.logger.printLogs();
@@ -83,11 +41,11 @@ describe('AudioTest', async function () {
 
     describe('setup', async function () {
       it('should open the meeting demo in two tabs', async function () {
-        test_window = await Window.existing(driverFactoryOne.driver, this.logger, 'TEST');
-        monitor_window = await Window.openNew(driverFactoryOne.driver, this.logger, 'MONITOR');
+        test_window = await Window.existing(this.driverFactoryOne.driver, this.logger, 'TEST');
+        monitor_window = await Window.openNew(this.driverFactoryOne.driver, this.logger, 'MONITOR');
 
-        await test_window.runCommands(async () => await pageOne.open(driverFactoryOne.url));
-        await monitor_window.runCommands(async () => await pageOne.open(driverFactoryOne.url));
+        await test_window.runCommands(async () => await this.pageOne.open(this.driverFactoryOne.url));
+        await monitor_window.runCommands(async () => await this.pageOne.open(this.driverFactoryOne.url));
       });
 
       it('should authenticate the user to the meeting', async function () {
@@ -95,67 +53,67 @@ describe('AudioTest', async function () {
         test_attendee_id = 'Test Attendee - ' + uuidv4().toString();
         monitor_attendee_id = 'Monitor Attendee - ' + uuidv4().toString();
 
-        await test_window.runCommands(async () => await pageOne.enterMeetingId(meetingId));
-        await monitor_window.runCommands(async () => await pageOne.enterMeetingId(meetingId));
+        await test_window.runCommands(async () => await this.pageOne.enterMeetingId(meetingId));
+        await monitor_window.runCommands(async () => await this.pageOne.enterMeetingId(meetingId));
 
         await test_window.runCommands(
-          async () => await pageOne.enterAttendeeName(test_attendee_id)
+          async () => await this.pageOne.enterAttendeeName(test_attendee_id)
         );
         await monitor_window.runCommands(
-          async () => await pageOne.enterAttendeeName(monitor_attendee_id)
+          async () => await this.pageOne.enterAttendeeName(monitor_attendee_id)
         );
 
         // TODO: Add region selection option if needed
         // await test_window.runCommands(async () => await page.selectRegion());
         // await monitor_window.runCommands(async () => await page.selectRegion());
 
-        await test_window.runCommands(async () => await pageOne.authenticate());
-        await monitor_window.runCommands(async () => await pageOne.authenticate());
+        await test_window.runCommands(async () => await this.pageOne.authenticate());
+        await monitor_window.runCommands(async () => await this.pageOne.authenticate());
       });
     });
 
     describe('user', async function () {
       it('should join the meeting', async function () {
-        await test_window.runCommands(async () => await pageOne.joinMeeting());
-        await monitor_window.runCommands(async () => await pageOne.joinMeeting());
+        await test_window.runCommands(async () => await this.pageOne.joinMeeting());
+        await monitor_window.runCommands(async () => await this.pageOne.joinMeeting());
       });
     });
 
     describe('meeting', async function () {
       it('should have two participants in the roster', async function () {
-        await test_window.runCommands(async () => await pageOne.rosterCheck(2));
-        await monitor_window.runCommands(async () => await pageOne.rosterCheck(2));
+        await test_window.runCommands(async () => await this.pageOne.rosterCheck(2));
+        await monitor_window.runCommands(async () => await this.pageOne.rosterCheck(2));
       });
     });
 
     describe('both attendee', async () => {
       it('should be muted at the start of the test', async function () {
-        await test_window.runCommands(async () => await pageOne.muteMicrophone());
-        await monitor_window.runCommands(async () => await pageOne.muteMicrophone());
+        await test_window.runCommands(async () => await this.pageOne.muteMicrophone());
+        await monitor_window.runCommands(async () => await this.pageOne.muteMicrophone());
       });
     });
 
     describe('test attendee', async function () {
       it('should play random tone', async function () {
-        await test_window.runCommands(async () => await pageOne.playRandomTone());
+        await test_window.runCommands(async () => await this.pageOne.playRandomTone());
       });
     });
 
     describe('monitor attendee', async function () {
       it('should check for random tone in the meeting', async function () {
-        await monitor_window.runCommands(async () => await pageOne.runAudioCheck('AUDIO_ON'));
+        await monitor_window.runCommands(async () => await this.pageOne.runAudioCheck('AUDIO_ON'));
       });
     });
 
     describe('test user', async function () {
       it('should stop playing random tone', async function () {
-        await test_window.runCommands(async () => await pageOne.stopPlayingRandomTone());
+        await test_window.runCommands(async () => await this.pageOne.stopPlayingRandomTone());
       });
     });
 
     describe('monitor user', async function () {
       it('should check for no random tone in the meeting', async function () {
-        await test_window.runCommands(async () => await pageOne.runAudioCheck('AUDIO_OFF'));
+        await test_window.runCommands(async () => await this.pageOne.runAudioCheck('AUDIO_OFF'));
       });
     });
   });
